Also try /usr/share/{NCINE_LINUX_PACKAGE}/Source when packaged on UNIX
This is tried before ~/.local/share/{NCINE_LINUX_PACKAGE}/Source, which
is slightly counter-intuitive but necessary. If you favour the user
directory, it will be automatically created on first launch and the
system-wide directory will never be considered again, even if the user
directory is still empty. The game shouldn't try to create the
system-wide directory because it wouldn't have permission to do so.